Moved main window class into its own file
