+++ CustomerOrganizer/CustomerOrganizer/CustomerOrganizer.py
@@ -1,25 +1,6 @@
-import tkinter as tk
+import MainWindow
+import tkinter
 
-class Application(tk.Frame):
-    def __init__(self, master = None):
-        super().__init__(master)
-        self.master = master
-        self.pack()
-        self.create_widgets()
-
-    def create_widgets(self):
-        self.hi_there = tk.Button(self)
-        self.hi_there["text"] = "Hello World\n(click me)"
-        self.hi_there["command"] = self.say_hi
-        self.hi_there.pack(side="top")
-
-        self.quit = tk.Button(self, text="QUIT", fg="red",
-                              command=self.master.destroy)
-        self.quit.pack(side="bottom")
-
-    def say_hi(self):
-        print("hi there, everyone!")
-
-root = tk.Tk()
-app = Application(master=root)
+root = tkinter.Tk()
+app = MainWindow.Application(master=root)
 app.mainloop()
