+++ CustomerOrganizer/CustomerOrganizer/CustomerOrganizer.pyproj
@@ -21,6 +21,9 @@
     <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="MainWindow.py">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="CustomerOrganizer.py" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets" />
